--> completed admin service layer business logic

+++ src/main/java/com/square/health/model/Post.java
@@ -24,7 +24,7 @@ public class Post {
     @Column(length = 65535, columnDefinition="TEXT")
     private String post; // need to add constraint Text
     private Long approvedBy;
-    private Byte isPublished;
+    private Integer isPublished;
     @Temporal(TemporalType.DATE)
     @Column(nullable = false, updatable = false)
     private Date createDate;
